move train to BaseTrainer
