+++ pytorch_lightning/core/lightning.py
@@ -19,6 +19,7 @@ from argparse import Namespace
 import collections
 import copy
 import inspect
+import os
 from pathlib import Path
 import re
 import tempfile
