+++ pytorch_lightning/trainer/base_trainer.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from abc import abstractmethod
+from abc import abstractclassmethod, abstractmethod
 from argparse import ArgumentParser, Namespace
 from copy import deepcopy
 import inspect
@@ -84,6 +84,20 @@ class BaseTrainer:
     #    Trainer Attributes   #
     #                         #
     ###########################
+
+    """
+    This class regroups the following Mixins:
+        - TrainerInternalLogicMixin:
+            Functions used by fit, train
+        - TrainerProperties
+        - TrainerCallbackHookMixin
+        - TrainerModelHooksMixin
+        - TrainerOptimizersMixin
+        - TrainerLoggingMixin
+        - TrainerTrainingTricksMixin
+        - TrainerDataLoadingMixin
+        - DeprecatedDistDeviceAttributes
+    """
 
     # model and datamodule
     model: LightningModule
@@ -193,18 +207,92 @@ class BaseTrainer:
     #                             #
     ###############################
 
+    @abstractclassmethod
     def fit(self, *_, **__):
-        raise NotImplementedError
+        pass
 
+    @abstractclassmethod
     def tune(self, *_, **__):
-        raise NotImplementedError
+        pass
 
+    @abstractclassmethod
     def test(self, *_, **__):
-        raise NotImplementedError
+        pass
 
-    ################################
-    #   TrainerInternalLogicMixin  #
-    ################################
+    ########################
+    #                      #
+    #    TRAIN FUNCTION    #
+    #                      #
+    ########################
+
+    def train(self):
+        self.run_sanity_check(self.get_model())
+
+        # set stage for logging
+        self.logger_connector.set_stage("train")
+
+        self.checkpoint_connector.has_trained = False
+
+        # enable train mode
+        model = self.get_model()
+        model.train()
+        torch.set_grad_enabled(True)
+
+        # reload data when needed
+        self.train_loop.reset_train_val_dataloaders(model)
+
+        # hook
+        self.train_loop.on_train_start()
+
+        try:
+            if self.train_loop.should_skip_training():
+                return
+            # run all epochs
+            for epoch in range(self.current_epoch, self.max_epochs):
+
+                # hook
+                self.train_loop.on_train_epoch_start(epoch)
+
+                with self.profiler.profile("run_training_epoch"):
+                    # run train epoch
+                    self.train_loop.run_training_epoch()
+
+                if self.max_steps and self.max_steps <= self.global_step:
+                    return
+
+                # update LR schedulers
+                self.optimizer_connector.update_learning_rates(interval='epoch')
+
+                # early stopping
+                met_min_epochs = epoch >= self.min_epochs - 1
+                met_min_steps = self.global_step >= self.min_steps if self.min_steps else True
+
+                if self.should_stop:
+                    if met_min_epochs and met_min_steps:
+                        return
+                    log.info(
+                        'Trainer was signaled to stop but required minimum epochs'
+                        f' ({self.min_epochs}) or minimum steps ({self.min_steps}) has'
+                        ' not been met. Training will continue...'
+                    )
+
+        except KeyboardInterrupt:
+            rank_zero_warn('Detected KeyboardInterrupt, attempting graceful shutdown...')
+
+            # user could press ctrl+c many times... only shutdown once
+            if not self.interrupted:
+                self.interrupted = True
+                self._state = TrainerState.INTERRUPTED
+                self.on_keyboard_interrupt()
+        finally:
+            # hook
+            self.train_loop.on_train_end()
+
+    ########################
+    #                      #
+    #    TEST FUNCTIONS    #
+    #                      #
+    ########################
 
     def run_evaluation(self, test_mode: bool = False, max_batches=None):
 
@@ -311,6 +399,12 @@ class BaseTrainer:
                         result[k] = v.cpu().item()
 
         return eval_loop_results
+
+    ###################################
+    #                                 #
+    #    TrainerInternalLogicMixin    #
+    #                                 #
+    ###################################
 
     def track_output_for_epoch_end(self, outputs, output):
         if output is not None:
