+++ pytorch_lightning/trainer/trainer.py
@@ -32,7 +32,6 @@ from pytorch_lightning.loggers import LightningLoggerBase
 from pytorch_lightning.plugins.plugin_connector import PluginConnector
 from pytorch_lightning.profiler import BaseProfiler
 from pytorch_lightning.trainer.base_trainer import BaseTrainer
-from pytorch_lightning.trainer.callback_hook import TrainerCallbackHookMixin
 from pytorch_lightning.trainer.configuration_validator import ConfigValidator
 from pytorch_lightning.trainer.connectors.callback_connector import CallbackConnector
 from pytorch_lightning.trainer.connectors.checkpoint_connector import CheckpointConnector
@@ -469,69 +468,6 @@ class Trainer(BaseTrainer):
         if self._state != TrainerState.INTERRUPTED:
             self._state = TrainerState.FINISHED
         return results or 1
-
-    def train(self):
-        self.run_sanity_check(self.get_model())
-
-        # set stage for logging
-        self.logger_connector.set_stage("train")
-
-        self.checkpoint_connector.has_trained = False
-
-        # enable train mode
-        model = self.get_model()
-        model.train()
-        torch.set_grad_enabled(True)
-
-        # reload data when needed
-        self.train_loop.reset_train_val_dataloaders(model)
-
-        # hook
-        self.train_loop.on_train_start()
-
-        try:
-            if self.train_loop.should_skip_training():
-                return
-            # run all epochs
-            for epoch in range(self.current_epoch, self.max_epochs):
-
-                # hook
-                self.train_loop.on_train_epoch_start(epoch)
-
-                with self.profiler.profile("run_training_epoch"):
-                    # run train epoch
-                    self.train_loop.run_training_epoch()
-
-                if self.max_steps and self.max_steps <= self.global_step:
-                    return
-
-                # update LR schedulers
-                self.optimizer_connector.update_learning_rates(interval='epoch')
-
-                # early stopping
-                met_min_epochs = epoch >= self.min_epochs - 1
-                met_min_steps = self.global_step >= self.min_steps if self.min_steps else True
-
-                if self.should_stop:
-                    if met_min_epochs and met_min_steps:
-                        return
-                    log.info(
-                        'Trainer was signaled to stop but required minimum epochs'
-                        f' ({self.min_epochs}) or minimum steps ({self.min_steps}) has'
-                        ' not been met. Training will continue...'
-                    )
-
-        except KeyboardInterrupt:
-            rank_zero_warn('Detected KeyboardInterrupt, attempting graceful shutdown...')
-
-            # user could press ctrl+c many times... only shutdown once
-            if not self.interrupted:
-                self.interrupted = True
-                self._state = TrainerState.INTERRUPTED
-                self.on_keyboard_interrupt()
-        finally:
-            # hook
-            self.train_loop.on_train_end()
 
     def test(
         self,
